feat: update to latest preview
