+++ package.json
@@ -33,7 +33,7 @@
     "@types/mdast": "^4.0.4",
     "@types/node": "^22.18.1",
     "@typescript-eslint/parser": "^8.51.0",
-    "astro": "^6.0.2",
+    "astro": "https://pkg.pr.new/astro@16149",
     "astro-eslint-parser": "^1.2.2",
     "astro-og-canvas": "^0.10.0",
     "canvaskit-wasm": "^0.40.0",
@@ -59,6 +59,7 @@
   "dependencies": {
     "@astrojs/check": "^0.9.7",
     "@astrojs/compiler-rs": "^0.1.6",
+    "@astrojs/markdown-remark": "https://pkg.pr.new/@astrojs/markdown-remark@16149",
     "@astrojs/mdx": "https://pkg.pr.new/@astrojs/mdx@16149?fdsfsd",
     "@astrojs/sitemap": "^3.7.1",
     "@astrojs/starlight": "^0.38.0",
@@ -68,9 +69,9 @@
     "jsdoc-api": "^9.3.5",
     "rehype-slug": "^6.0.0",
     "remark-smartypants": "^3.0.2",
+    "satteri": "^0.2.5",
     "sharp": "^0.34.3",
-    "starlight-llms-txt": "^0.6.0",
-    "tryckeri": "^0.1.23"
+    "starlight-llms-txt": "^0.6.0"
   },
   "packageManager": "pnpm@10.27.0",
   "pnpm": {
